backend/requirements.txt
Django==4.2
djangorestframework==3.14.0
psycopg2-binary==2.9.6
djangorestframework-simplejwt==5.2.2
django-cors-headers==4.0.0
Pillow==10.0.0
python-decouple==3.8
django-filter==23.2
drf-yasg==1.21.5